fix(input): let libc type the FIONREAD request (musl declares it c_int)
glibc 的 ioctl request 形参是 c_ulong,musl 是 c_int;libc 已把
FIONREAD 按平台定义为对应的 Ioctl 类型,直接传即可,显式 as c_ulong
反而在 musl check 下 E0308。linux-gnu 行为不变。

+++ src/input.rs
@@ -123,12 +123,10 @@ fn fd0_has_unread_bytes() -> bool {
             }
             _ => {
                 let mut pending: libc::c_int = 0;
-                if libc::ioctl(
-                    libc::STDIN_FILENO,
-                    libc::FIONREAD as libc::c_ulong,
-                    &mut pending,
-                ) == 0
-                {
+                // FIONREAD is typed as the platform's ioctl request type in
+                // libc (c_ulong under glibc, c_int under musl): no cast, or
+                // the musl build breaks.
+                if libc::ioctl(libc::STDIN_FILENO, libc::FIONREAD, &mut pending) == 0 {
                     pending > 0
                 } else {
                     let mut fds = [libc::pollfd {
